Struct halo2curves::secp256k1::Fp

pub struct Fp(_);

This represents an element of $\mathbb{F}_p$ where

p = 0xfffffffffffffffffffffffffffffffffffffffffffffffffffffffefffffc2f

is the base field of the secp256k1 curve.

Implementations

impl Fp

pub const fn zero() -> Fp

Returns zero, the additive identity.

pub const fn one() -> Fp

Returns one, the multiplicative identity.

pub const fn from_raw(val: [u64; 4]) -> Self

Converts from an integer represented in little endian into its (congruent) $field representation.

pub fn from_bytes(bytes: &[u8; 32]) -> CtOption<Fp>

Attempts to convert a little-endian byte representation of a scalar into a Fr, failing if the input is not canonical.

pub fn to_bytes(&self) -> [u8; 32]

Converts an element of Fr into a byte representation in little-endian byte order.

impl Fp

pub const fn add(&self, rhs: &Self) -> Self

Adds rhs to self, returning the result.

pub const fn mul(&self, rhs: &Self) -> Fp

Multiplies rhs by self, returning the result.

impl Fp

pub const fn double(&self) -> Fp

Doubles this field element.

pub const fn square(&self) -> Fp

Squares this element.

pub const fn sub(&self, rhs: &Self) -> Self

Subtracts rhs from self, returning the result.

pub const fn neg(&self) -> Self

Negates self.

impl Fp

pub const fn size() -> usize