Trait halo2curves::FieldExt

pub trait FieldExt: SqrtRatio + From<bool> + Ord + Group<Scalar = Self> {
    const MODULUS: &'static str;
    const ROOT_OF_UNITY_INV: Self;
    const DELTA: Self;
    const TWO_INV: Self;
    const ZETA: Self;

    fn from_u128(v: u128) -> Self;
    fn from_bytes_wide(bytes: &[u8; 64]) -> Self;
    fn get_lower_128(&self) -> u128;

    fn pow(&self, by: &[u64; 4]) -> Self { ... }
}

This trait is a common interface for dealing with elements of a finite field.

Required Associated Constants

const MODULUS: &'static str

Modulus of the field written as a string for display purposes

const ROOT_OF_UNITY_INV: Self

Inverse of PrimeField::root_of_unity()

const DELTA: Self

Generator of the $t-order$ multiplicative subgroup

const TWO_INV: Self

Inverse of $2$ in the field.

const ZETA: Self

Element of multiplicative order $3$.

Required Methods

fn from_u128(v: u128) -> Self

Obtains a field element congruent to the integer v.

fn from_bytes_wide(bytes: &[u8; 64]) -> Self

Obtains a field element that is congruent to the provided little endian byte representation of an integer.

fn get_lower_128(&self) -> u128

Gets the lower 128 bits of this field element when expressed canonically.

Provided Methods

fn pow(&self, by: &[u64; 4]) -> Self

Exponentiates self by by, where by is a little-endian order integer exponent.