Crate halo2curves

Re-exports

pub extern crate group;

Modules

bn256

pairing

pasta

secp256k1

serde

Macros

batch_add

field_arithmetic

field_common

field_specific

impl_add_binop_specify_output

impl_binops_additive

impl_binops_additive_specify_output

impl_binops_multiplicative

impl_binops_multiplicative_mixed

impl_sub_binop_specify_output

new_curve_impl

Structs

Coordinates

The affine coordinates of a point on an elliptic curve.

Traits

CurveAffine

This trait is the affine counterpart to Curve and is used for serialization, storage in memory, and inspection of $x$ and $y$ coordinates.

CurveAffineExt

CurveExt

This trait is a common interface for dealing with elements of an elliptic curve group in a “projective” form, where that arithmetic is usually more efficient.

FieldExt

This trait is a common interface for dealing with elements of a finite field.

Group

This represents an element of a group with basic operations that can be performed. This allows an FFT implementation (for example) to operate generically over either a field or elliptic curve group.